requirements.txt
fastapi>=0.100.0
uvicorn>=0.30.0
pydantic>=2.0.0
PyMuPDF>=1.24.0
pdfplumber>=0.10.0
python-docx>=1.0.0
httpx>=0.25.0
aiofiles>=23.0.0
python-multipart>=0.0.6
regex>=2024.0.0
chardet>=5.0.0
trafilatura>=1.9.0
openpyxl>=3.1.2
playwright>=1.45.0
